requirements.txt
chaindl
pandas
requests
python-dotenv